Apply Black to standardize code styling (#2614)
This uses the [black](https://github.com/ambv/black) code styler to manage our style. 

1.  Apply `black` to the entire codebase
2.  Add `black` to our Travis CI
3.  Add a pre-commit hook to make this automatic, 
    should people use `pre-commit install` at the base of this repository

The only concern I have about this is that it screws with `git blame`.  

To future maintainers that come across this commit, 
I recommend heading back one more commit `HEAD^` 
and then using `git blame` again.

+++ distributed/protocol/tests/test_h5py.py
@@ -3,7 +3,7 @@ import traceback
 
 import pytest
 
-h5py = pytest.importorskip('h5py')
+h5py = pytest.importorskip("h5py")
 
 from distributed.protocol import deserialize, serialize
 
@@ -33,39 +33,39 @@ def silence_h5py_issue775(func):
 @silence_h5py_issue775
 def test_serialize_deserialize_file():
     with tmpfile() as fn:
-        with h5py.File(fn, mode='a') as f:
-            f.create_dataset('/x', shape=(2, 2), dtype='i4')
-        with h5py.File(fn, mode='r') as f:
+        with h5py.File(fn, mode="a") as f:
+            f.create_dataset("/x", shape=(2, 2), dtype="i4")
+        with h5py.File(fn, mode="r") as f:
             g = deserialize(*serialize(f))
             assert f.filename == g.filename
             assert isinstance(g, h5py.File)
             assert f.mode == g.mode
 
-            assert g['x'].shape == (2, 2)
+            assert g["x"].shape == (2, 2)
 
 
 @silence_h5py_issue775
 def test_serialize_deserialize_group():
     with tmpfile() as fn:
-        with h5py.File(fn, mode='a') as f:
-            f.create_dataset('/group1/group2/x', shape=(2, 2), dtype='i4')
-        with h5py.File(fn, mode='r') as f:
-            group = f['/group1/group2']
+        with h5py.File(fn, mode="a") as f:
+            f.create_dataset("/group1/group2/x", shape=(2, 2), dtype="i4")
+        with h5py.File(fn, mode="r") as f:
+            group = f["/group1/group2"]
             group2 = deserialize(*serialize(group))
 
             assert isinstance(group2, h5py.Group)
             assert group.file.filename == group2.file.filename
 
-            assert group2['x'].shape == (2, 2)
+            assert group2["x"].shape == (2, 2)
 
 
 @silence_h5py_issue775
 def test_serialize_deserialize_dataset():
     with tmpfile() as fn:
-        with h5py.File(fn, mode='a') as f:
-            x = f.create_dataset('/group1/group2/x', shape=(2, 2), dtype='i4')
-        with h5py.File(fn, mode='r') as f:
-            x = f['group1/group2/x']
+        with h5py.File(fn, mode="a") as f:
+            x = f.create_dataset("/group1/group2/x", shape=(2, 2), dtype="i4")
+        with h5py.File(fn, mode="r") as f:
+            x = f["group1/group2/x"]
             y = deserialize(*serialize(x))
             assert isinstance(y, h5py.Dataset)
             assert x.name == y.name
@@ -76,8 +76,8 @@ def test_serialize_deserialize_dataset():
 @silence_h5py_issue775
 def test_raise_error_on_serialize_write_permissions():
     with tmpfile() as fn:
-        with h5py.File(fn, mode='a') as f:
-            x = f.create_dataset('/x', shape=(2, 2), dtype='i4')
+        with h5py.File(fn, mode="a") as f:
+            x = f.create_dataset("/x", shape=(2, 2), dtype="i4")
             f.flush()
             with pytest.raises(TypeError):
                 deserialize(*serialize(x))
@@ -95,14 +95,14 @@ import dask.array as da
 @gen_cluster(client=True)
 def test_h5py_serialize(c, s, a, b):
     from dask.utils import SerializableLock
-    lock = SerializableLock('hdf5')
+
+    lock = SerializableLock("hdf5")
     with tmpfile() as fn:
-        with h5py.File(fn, mode='a') as f:
-            x = f.create_dataset('/group/x', shape=(4,), dtype='i4',
-                                 chunks=(2,))
+        with h5py.File(fn, mode="a") as f:
+            x = f.create_dataset("/group/x", shape=(4,), dtype="i4", chunks=(2,))
             x[:] = [1, 2, 3, 4]
-        with h5py.File(fn, mode='r') as f:
-            dset = f['/group/x']
+        with h5py.File(fn, mode="r") as f:
+            dset = f["/group/x"]
             x = da.from_array(dset, chunks=dset.chunks, lock=lock)
             y = c.compute(x)
             y = yield y
@@ -112,12 +112,11 @@ def test_h5py_serialize(c, s, a, b):
 @gen_cluster(client=True)
 def test_h5py_serialize_2(c, s, a, b):
     with tmpfile() as fn:
-        with h5py.File(fn, mode='a') as f:
-            x = f.create_dataset('/group/x', shape=(12,), dtype='i4',
-                                 chunks=(4,))
+        with h5py.File(fn, mode="a") as f:
+            x = f.create_dataset("/group/x", shape=(12,), dtype="i4", chunks=(4,))
             x[:] = [1, 2, 3, 4] * 3
-        with h5py.File(fn, mode='r') as f:
-            dset = f['/group/x']
+        with h5py.File(fn, mode="r") as f:
+            dset = f["/group/x"]
             x = da.from_array(dset, chunks=(3,))
             y = c.compute(x.sum())
             y = yield y
