+++ distributed/tests/test_batched.py
@@ -28,7 +28,7 @@ class EchoServer(object):
                 return
 
     def listen(self):
-        listener = listen('', self.handle_comm)
+        listener = listen("", self.handle_comm)
         listener.start()
         self.address = listener.contact_address
         self.stop = listener.stop
@@ -57,17 +57,17 @@ def test_BatchedSend():
 
         yield gen.sleep(0.020)
 
-        b.send('hello')
-        b.send('hello')
-        b.send('world')
+        b.send("hello")
+        b.send("hello")
+        b.send("world")
         yield gen.sleep(0.020)
-        b.send('HELLO')
-        b.send('HELLO')
+        b.send("HELLO")
+        b.send("HELLO")
 
         result = yield comm.read()
-        assert result == ('hello', 'hello', 'world')
+        assert result == ("hello", "hello", "world")
         result = yield comm.read()
-        assert result == ('HELLO', 'HELLO')
+        assert result == ("HELLO", "HELLO")
 
         assert b.byte_count > 1
 
@@ -79,12 +79,12 @@ def test_send_before_start():
 
         b = BatchedSend(interval=10)
 
-        b.send('hello')
-        b.send('world')
+        b.send("hello")
+        b.send("world")
 
         b.start(comm)
         result = yield comm.read()
-        assert result == ('hello', 'world')
+        assert result == ("hello", "world")
 
 
 @gen_test()
@@ -95,12 +95,12 @@ def test_send_after_stream_start():
         b = BatchedSend(interval=10)
 
         b.start(comm)
-        b.send('hello')
-        b.send('world')
+        b.send("hello")
+        b.send("world")
         result = yield comm.read()
         if len(result) < 2:
             result += yield comm.read()
-        assert result == ('hello', 'world')
+        assert result == ("hello", "world")
 
 
 @gen_test()
@@ -112,8 +112,8 @@ def test_send_before_close():
         b.start(comm)
 
         cnt = int(e.count)
-        b.send('hello')
-        yield b.close()         # close immediately after sending
+        b.send("hello")
+        yield b.close()  # close immediately after sending
         assert not b.buffer
 
         start = time()
@@ -122,7 +122,7 @@ def test_send_before_close():
             assert time() < start + 5
 
         with pytest.raises(CommClosedError):
-            b.send('123')
+            b.send("123")
 
 
 @gen_test()
@@ -137,8 +137,8 @@ def test_close_closed():
         comm.close()  # external closing
 
         yield b.close()
-        assert 'closed' in repr(b)
-        assert 'closed' in str(b)
+        assert "closed" in repr(b)
+        assert "closed" in str(b)
 
 
 @gen_test()
@@ -191,18 +191,19 @@ def test_stress():
 @gen.coroutine
 def run_traffic_jam(nsends, nbytes):
     # This test eats `nsends * nbytes` bytes in RAM
-    np = pytest.importorskip('numpy')
+    np = pytest.importorskip("numpy")
     from distributed.protocol import to_serialize
-    data = bytes(np.random.randint(0, 255, size=(nbytes,)).astype('u1').data)
+
+    data = bytes(np.random.randint(0, 255, size=(nbytes,)).astype("u1").data)
     with echo_server() as e:
         comm = yield connect(e.address)
 
         b = BatchedSend(interval=0.01)
         b.start(comm)
 
-        msg = {'x': to_serialize(data)}
+        msg = {"x": to_serialize(data)}
         for i in range(nsends):
-            b.send(assoc(msg, 'i', i))
+            b.send(assoc(msg, "i", i))
             if np.random.random() > 0.5:
                 yield gen.sleep(0.001)
 
@@ -214,7 +215,7 @@ def run_traffic_jam(nsends, nbytes):
             # loses some of our messages
             L = yield gen.with_timeout(timedelta(seconds=5), comm.read())
             count += 1
-            results.extend(r['i'] for r in L)
+            results.extend(r["i"] for r in L)
 
         assert count == b.batch_count == e.count
         assert b.message_count == nsends
@@ -241,25 +242,25 @@ def test_serializers():
     with echo_server() as e:
         comm = yield connect(e.address)
 
-        b = BatchedSend(interval='10ms', serializers=['msgpack'])
+        b = BatchedSend(interval="10ms", serializers=["msgpack"])
         b.start(comm)
 
-        b.send({'x': to_serialize(123)})
-        b.send({'x': to_serialize('hello')})
+        b.send({"x": to_serialize(123)})
+        b.send({"x": to_serialize("hello")})
         yield gen.sleep(0.100)
 
-        b.send({'x': to_serialize(lambda x: x + 1)})
+        b.send({"x": to_serialize(lambda x: x + 1)})
 
-        with captured_logger('distributed.protocol') as sio:
+        with captured_logger("distributed.protocol") as sio:
             yield gen.sleep(0.100)
 
         value = sio.getvalue()
-        assert 'serialize' in value
-        assert 'type' in value
-        assert 'function' in value
+        assert "serialize" in value
+        assert "type" in value
+        assert "function" in value
 
         msg = yield comm.read()
-        assert list(msg) == [{'x': 123}, {'x': 'hello'}]
+        assert list(msg) == [{"x": 123}, {"x": "hello"}]
 
         with pytest.raises(gen.TimeoutError):
             msg = yield gen.with_timeout(timedelta(milliseconds=100), comm.read())
